test: add investigation merge and routing unit tests
