+++ tests/nodes/test_parallel_investigate.py
@@ -1,10 +1,36 @@
 """Tests for parallel investigate hypothesis routing and merging."""
 
+from __future__ import annotations
+
+from typing import Any
+
 from langgraph.constants import Send
 
+from app.nodes.investigate.execution.execute_actions import ActionExecutionResult
+from app.nodes.investigate import parallel as parallel_module
 from app.nodes.investigate.parallel import node_investigate_hypothesis
 from app.pipeline.routing import distribute_hypotheses
 from app.state.factory import make_initial_state
+
+
+class _TrackerStub:
+    def start(self, _name: str, _message: str) -> None:
+        return
+
+    def complete(
+        self,
+        _name: str,
+        *,
+        fields_updated: list[str],
+        message: str,
+    ) -> None:
+        assert fields_updated is not None
+        assert message is not None
+
+
+class _ActionStub:
+    def __init__(self, name: str) -> None:
+        self.name = name
 
 
 def test_distribute_hypotheses_with_actions():
@@ -50,3 +76,71 @@ def test_node_investigate_hypothesis_unknown_action():
 
     result = node_investigate_hypothesis(state)
     assert result == {"hypothesis_results": []}
+
+
+def test_node_investigate_hypothesis_success_serializes_result(monkeypatch: Any) -> None:
+    """Successful execution serializes action name, data, and success flag."""
+    action = _ActionStub("query_alertmanager")
+    execution_result = ActionExecutionResult(
+        action_name=action.name,
+        success=True,
+        data={"alerts": [{"labels": {"alertname": "HighErrorRate"}}]},
+        error=None,
+    )
+
+    monkeypatch.setattr(parallel_module, "get_tracker", _TrackerStub)
+    monkeypatch.setattr(parallel_module, "get_available_actions", lambda: [action])
+    monkeypatch.setattr(
+        parallel_module,
+        "execute_actions",
+        lambda _names, _actions, _sources: {action.name: execution_result},
+    )
+
+    state = dict(make_initial_state("HighErrorRate", "checkout", "critical"))
+    state["action_to_run"] = action.name
+    state["available_sources"] = {"alertmanager": {"url": "http://localhost:9093"}}
+
+    assert parallel_module.node_investigate_hypothesis(state) == {
+        "hypothesis_results": [
+            {
+                "action_name": action.name,
+                "success": True,
+                "data": {"alerts": [{"labels": {"alertname": "HighErrorRate"}}]},
+                "error": None,
+            }
+        ]
+    }
+
+
+def test_node_investigate_hypothesis_failure_serializes_result(monkeypatch: Any) -> None:
+    """Failed execution propagates error string into the serialized result."""
+    action = _ActionStub("query_alertmanager")
+    execution_result = ActionExecutionResult(
+        action_name=action.name,
+        success=False,
+        data={"available": False},
+        error="connection refused",
+    )
+
+    monkeypatch.setattr(parallel_module, "get_tracker", _TrackerStub)
+    monkeypatch.setattr(parallel_module, "get_available_actions", lambda: [action])
+    monkeypatch.setattr(
+        parallel_module,
+        "execute_actions",
+        lambda _names, _actions, _sources: {action.name: execution_result},
+    )
+
+    state = dict(make_initial_state("HighErrorRate", "checkout", "critical"))
+    state["action_to_run"] = action.name
+
+    assert parallel_module.node_investigate_hypothesis(state) == {
+        "hypothesis_results": [
+            {
+                "action_name": action.name,
+                "success": False,
+                "data": {"available": False},
+                "error": "connection refused",
+            }
+        ]
+    }
+
